Abstract tab_label logic
This commit abstracts the logic that checks if a tab in the tabline has a
modified buffer or not to a function (`get_tab_label`) which does
that check, and returns the corresponding tab_label to be used in the
tabline. It also handles some errors that would have previously
caused gnvim to crash.