Refactor previous changes
Some notable changes in this commit include the additoin of the
'CompletionItemKind' enum to nvim_bridge.rs per vhakulinen's
suggestion, in addition to changing the 'CompletionItemWidgetWrap::create'
function to take a reference to the other items in the popupmenu.
That function then checks whether or not any other items in the
pmenu have icons to determine if it should use an icon
for the created 'CompletionItemWidgetWrap' object if it's kind is
unknown.

I have also refactored `get_icon_pixbuf` and `get_icon_name_for_kind`
to take a reference to a 'CompletionItemKind' instead of a 'str'.