Update versions of dependencies
Updates Cargo.toml and Cargo.lock to
use the ~latest versions of dependencies.

With the new GTK version, a few places
needed to be updated to use `Some` instead
of just a simple value. Also, many places
with `None` it needed to be modified to be more
explicit, for example `None::<&gtk::Adjustment>`
instead of a simple `None`. This commit makes
those changes.