Implement smooth scrolling support
Scrolling with the touchpad had no effect. It seemed scrolling events
were purely ignored. As explained in #70, this is due to smooth
scrolling. The problem with enabling smooth scrolling is that it
generates a lot of scrolling events, resulting in overly fast scroll
without further processing.

The solution here consists in hooking on the "start scroll" event to
reset a scrolling counter (this requires to bump the GTK version). When
the scrolling is hapening, this counter is incremented with the
scrolling deltas.  When it reaches -1.0 or +1.0 (which is the value that
"normal scrolling" sends), we trigger a scrolling event (down or up,
depending on the sign) and reset the counter.

Overall, it enables scrolling in gnvim with the touchpad.

This probably fixes #70.