ui: implement external windows
Implement external windows. This results in various changes.

First, the neovim rpc client gets wrapped into a gobject so its easier
to pass around (through binding). This is mainly required for the
external window, and allow it to be resized.

Second, the mouse input is changed to be grid based instead of always
targeting the root grid. This is to allow mouse input on the external
windows.

And third, the external window itself. The grid that gets placed in the
external window owns it (the external window).