nvim-rs: flatten the rpc client
Introduce a rpc `Caller` for which the neovim api is implemented for.
This removes the need to have a explicit client implementation provided
from the nvim-rs crate, and allows the UI to handle more of the client
logic.

In otherwords, make the nvim-rs crate responsibility more limited to the
RPC things.

Additionally, use the message structs for serialization.