ui: subclass gtk::Application for our application
Subclass gtk::Application to take better control of the application
instance. Now we support a "primary" instance model that gtk comes with.
`--new` flag allows to launch a new instance of gnvim.

Transition from clap to gtk::Application's command line parsing. Gtk,
which is our main dependency, already provides the functionaly we
need/use so might as well use it and avoid a extra dependency.