ui: create grids on first use
As noted by one of the neovim core devs, the transmitted UI state is
"promised to be consistent on flush". The docs for `grid_resize` docs
mention that a "new grid is being created" on that event, but it doesn't
apparently mean that `grid_resize` event would be the first event to
"declare" a new grid - only that the grid should be resized (allocated)
to the given size.