ui: reposition grid in `grid_resize`
Grid's positions are anchored. To keep the anchored, we need to
"reposition" them when their size changes.

This fix removes the clamping of floating windows. I don't remember it
(the clamping) ever happening. It also moves the handling of the
grid/window size from nvim to the GUI, and we don't have any logic for
that.